Shamirs Secret Sharing GUI changes:
- mode selection was removec. graphical mode is the only one that can be
used from now on.
- default values are set: n=4, t=3, p=23, s=8
- indices are super /subset
- coefficient dialog fixed
- Text in group information added with more information
- reset button resets everything except of n, t, s, p.
