+++ org.jcryptool.visual.secretsharing/src/org/jcryptool/visual/secretsharing/views/messages.properties
@@ -1,3 +1,5 @@
+lagrange_formular = \u03A3 s\u1d62 \u2022 \u03A0 ( x - x\u1d63 ) \u2022 ( x\u1d62 - x\u1d63 )\u207B\u00B9 mod p
+ShamirsCompositeGraphical_title=Shamir's Secret Sharing
 SSSConstants_Coefficient=Coefficient:
 SSSConstants_Compute_Share_Button=Compute shares
 SSSConstants_Concerned_Persons=Number of concerned persons n:
@@ -20,11 +22,11 @@ SSSConstants_Graphical_Mode=Graphical
 SSSConstants_Info_Group=Info
 SSSConstants_Modul_info=Module p (p \u2208 prime, p > n):
 SSSConstants_Numerical_Mode=Numerical
-SSSConstants_Polynom_Equal=The polynomials P(x) and P'(x) are equal. The secret s is the coefficient a\u2080.
+SSSConstants_Polynom_Equal=The polynomials P(x) and P''(x) are equal. The secret s is the coefficient a\u2080 = {0}.
 SSSConstants_Polynom_Info=Polynomial: s + a\u2081x + a\u2082x\u00B2 + ... + a\u1D63 x^r \t   (r := t-1)
-SSSConstants_Polynom_Not_Equal=The polynomials P(x) and P'(x) are not equal. It is not possible to reconstruct the polynomial P(x), add more points for the reconstruction. The reconstructed polynomial is false.
+SSSConstants_Polynom_Not_Equal=The reconstruction failed because only {0} instead of {1} shares were selected. Thus, the searched polynomial P (x) (blue) and the found P '(x) (red) are not equal. Add more points (shares) for the reconstruction.
 SSSConstants_Reconstrct=Reconstruct
-SSSConstants_Reconstruct_Info=In order to reconstruct the polynomial any t points will be enough. We use the Lagrange interpolation formula to reconstruct the polynomial:
+Reconstruct_Info=In order to reconstruct the polynomial any t points will be enough. We use the Lagrange interpolation formula to reconstruct the polynomial:
 SSSConstants_reconstruct_Person=Number of persons for reconstruction t:
 SSSConstants_Reconstruction_Group=Reconstruction
 SSSConstants_Secret_Info=Secret s (s < p):
@@ -33,7 +35,7 @@ SSSConstants_Select_All_Button=Select all
 SSSConstants_Select_Modus=Mode
 SSSConstants_Select_Parameter=Parameter
 SSSConstants_Settings=Settings
-SSSConstants_Title_Info=Shamir's Secret Sharing is a secret sharing algorithm for distributing a secret. The secret is divided into parts, giving each participant its own unique\npart where some of the parts or all of them are needed in order to reconstruct the secret. The secret is reconstructed with the Lagrange polynomial.\n
+SSSConstants_Title_Info=Shamir's Secret Sharing is a secret sharing algorithm for distributing a secret. The secret is divided into parts, giving each participant its own unique part where some of the parts or all of them are needed in order to reconstruct the secret. The secret is reconstructed with the Lagrange polynomial.
 SSSConstants_Title_Info_Formula=The Lagrange formula for the reconstruction is:
 SSSConstants_Zoom=Zoom:
 SSSConstants_Reset=Reset
