+++ org.jcryptool.visual.secretsharing/src/org/jcryptool/visual/secretsharing/views/messages_de.properties
@@ -1,3 +1,5 @@
+lagrange_formular = \u03A3 s\u1d62 \u2022 \u03A0 ( x - x\u1d63 ) \u2022 ( x\u1d62 - x\u1d63 )\u207B\u00B9 mod p
+ShamirsCompositeGraphical_title=Shamir's Secret Sharing
 SSSConstants_Coefficient=Koeffizient:
 SSSConstants_Compute_Share_Button=Shares berechnen
 SSSConstants_Concerned_Persons=Anzahl der beteiligten Personen n:
@@ -20,11 +22,11 @@ SSSConstants_Graphical_Mode=Graphisch
 SSSConstants_Info_Group=Information
 SSSConstants_Modul_info=Modul p (p \u2208 prim, p > n):
 SSSConstants_Numerical_Mode=Numerisch
-SSSConstants_Polynom_Equal=Die Polynome P(x) und P'(x) sind gleich. Das Geheimnis s ist der Koeffizient a\u2080.
+SSSConstants_Polynom_Equal=Die Polynome P(x) und P''(x) sind gleich. Das Geheimnis s ist der Koeffizient a\u2080 = {0}.
 SSSConstants_Polynom_Info=Polynom: s + a\u2081x + a\u2082x\u00B2 + ... + a\u1D63 x^r \t   (r := t-1)
-SSSConstants_Polynom_Not_Equal=Die Polynome P(x) und P'(x) sind nicht gleich. Es ist nicht m\u00f6glich das Polynom P(x) zu rekonstruieren. F\u00fcge mehr Punkte f\u00fcr die Rekonstruktion hinzu. Die Rekonstruktion ist falsch.
+SSSConstants_Polynom_Not_Equal=Die Rekonstruktion ist fehlgeschlagen, weil nur {0} statt {1} Shares ausgew\u00e4hlt wurden. Somit sind das gesuchte Polynom P(x) (blau) und das gefundenen P'(x) (rot) nicht gleich. F\u00fcge mehr Punkte (Shares) f\u00fcr die Rekonstruktion hinzu.
 SSSConstants_Reconstrct=Rekonstruieren
-SSSConstants_Reconstruct_Info=Um das Polynom rekonstruieren zu k\u00f6nnen gen\u00fcgen beliebige t Punkte. Wir benutzen die Lagrange-Interpolations Formel, um das Polynom zu rekonstruieren:
+Reconstruct_Info=Um das Polynom rekonstruieren zu k\u00f6nnen gen\u00fcgen beliebige t Punkte. Wir benutzen die Lagrange-Interpolations Formel, um das Polynom zu rekonstruieren:
 SSSConstants_reconstruct_Person=Anzahl Personen f\u00fcr Rekonstruktion t:
 SSSConstants_Reconstruction_Group=Rekonstruktion
 SSSConstants_Secret_Info=Geheimnis s (s < p):
@@ -33,7 +35,7 @@ SSSConstants_Select_All_Button=Alle w\u00e4hlen
 SSSConstants_Select_Modus=Modus
 SSSConstants_Select_Parameter=Parameter
 SSSConstants_Settings=Einstellungen
-SSSConstants_Title_Info=Shamir's Secret Sharing ist ein Secret Sharing Algorithmus, um ein Geheimnis zu verteilen. Das Geheimnis wird so verteilt, dass jede Person seinen\neigenen eindeutigen Teil bekommt. Einige dieser Teile werden ben\u00f6tigt, um das Geheimnis zu rekonstruieren. Das Geheimnis wird mit Hilfe\nder Lagrange Interpolation berechnet.
-SSSConstants_Title_Info_Formula=Die Lagrange Formel f\u00fcr die Rekonstruktion lautet:
+SSSConstants_Title_Info=Shamir's Secret Sharing ist ein Secret-Sharing-Algorithmus, um ein Geheimnis zu verteilen. Das Geheimnis wird so verteilt, dass jede Person seinen eigenen eindeutigen Teil (Share) bekommt. Einige dieser Teile werden ben\u00f6tigt, um das Geheimnis zu rekonstruieren. Das Geheimnis wird mit Hilfe der Lagrange-Interpolation berechnet.
+SSSConstants_Title_Info_Formula=Die Lagrange-Formel f\u00fcr die Rekonstruktion lautet:
 SSSConstants_Zoom=Zoom:
 SSSConstants_Reset=Zur\u00fccksetzen
